+++ build.gradle
@@ -9,7 +9,7 @@ buildscript {
         maven { url "https://plugins.gradle.org/m2/" }
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:${libraryVersion.gradle}"
+        classpath 'com.android.tools.build:gradle:3.1.4'
         classpath "com.github.dcendents:android-maven-gradle-plugin:${libraryVersion.mavenPlugin}"
         classpath "com.vanniktech:gradle-android-junit-jacoco-plugin:${libraryVersion.jacoco}"
         classpath "gradle.plugin.io.gitlab.arturbosch:detekt-gradle-plugin:${libraryVersion.detekt}"
