+++ codecov.yml
@@ -22,7 +22,4 @@ parsers:
       method: no
       macro: no
 
-comment:
-  layout: "reach, diff, flags, files, footer"
-  behavior: default
-  require_changes: no
+comment: off
