+++ fountain-core/src/main/java/com/xmartlabs/fountain/adapter/NetworkDataSourceAdapter.kt
@@ -37,53 +37,3 @@ interface NetworkDataSourceAdapter<T> {
 
   val pageFetcher: PageFetcher<T>
 }
-
-/**
- * Provides a [NetworkDataSourceAdapter] implementation of a [ListResponseWithEntityCount] response.
- * It is used when the service returns the entity count in the response.
- */
-internal class NetworkDataSourceWithTotalEntityCountAdapter<T, R : ListResponseWithEntityCount<T>>(
-    pageFetcher: PageFetcher<R>,
-    firstPage: Int = FountainConstants.DEFAULT_FIRST_PAGE
-) : NetworkDataSourceAdapter<ListResponse<T>> {
-  private val knownSizeResponseManager = KnownSizeResponseManager(firstPage)
-
-  override fun canFetch(page: Int, pageSize: Int) = knownSizeResponseManager.canFetch(page, pageSize)
-
-  override val pageFetcher = object : PageFetcher<ListResponse<T>> {
-    override fun fetchPage(page: Int, pageSize: Int, networkResultListener: NetworkResultListener<ListResponse<T>>) =
-        pageFetcher.fetchPage(page, pageSize, object : NetworkResultListener<R> {
-          override fun onSuccess(response: R) {
-            knownSizeResponseManager.onTotalEntityResponseArrived(response)
-            networkResultListener.onSuccess(response)
-          }
-
-          override fun onError(t: Throwable) = networkResultListener.onError(t)
-        })
-  }
-}
-
-/**
- * Provides a [NetworkDataSourceAdapter] implementation of a [ListResponseWithPageCount] response.
- * It is used when the service returns the page count in the response.
- */
-internal class NetworkDataSourceWithTotalPageCountAdapter<T, R : ListResponseWithPageCount<T>>(
-    pageFetcher: PageFetcher<R>,
-    firstPage: Int = FountainConstants.DEFAULT_FIRST_PAGE
-) : NetworkDataSourceAdapter<ListResponse<T>> {
-  private val knownSizeResponseManager = KnownSizeResponseManager(firstPage)
-
-  override fun canFetch(page: Int, pageSize: Int) = knownSizeResponseManager.canFetch(page, pageSize)
-
-  override val pageFetcher = object : PageFetcher<ListResponse<T>> {
-    override fun fetchPage(page: Int, pageSize: Int, networkResultListener: NetworkResultListener<ListResponse<T>>) =
-        pageFetcher.fetchPage(page, pageSize, object : NetworkResultListener<R> {
-          override fun onSuccess(response: R) {
-            knownSizeResponseManager.onTotalPageCountResponseArrived(pageSize, response)
-            networkResultListener.onSuccess(response)
-          }
-
-          override fun onError(t: Throwable) = networkResultListener.onError(t)
-        })
-  }
-}
