+++ fountain-core/src/test/java/com/xmartlabs/fountain/networkstate/NetworkStatusUnitTest.kt
@@ -5,9 +5,9 @@ import com.xmartlabs.fountain.ListResponse
 import com.xmartlabs.fountain.Listing
 import com.xmartlabs.fountain.NetworkState
 import com.xmartlabs.fountain.common.IntMockedListingCreator
-import com.xmartlabs.fountain.common.MockedNetworkDataSourceAdapter
-import com.xmartlabs.fountain.common.extensions.mockLifecycleEvents
-import com.xmartlabs.fountain.common.extensions.sendPageResponse
+import com.xmartlabs.fountain.testutils.MockedNetworkDataSourceAdapter
+import com.xmartlabs.fountain.testutils.extensions.mockLifecycleEvents
+import com.xmartlabs.fountain.testutils.extensions.sendPageResponse
 import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
