+++ fountain-core/src/test/java/com/xmartlabs/fountain/pagefetcher/totalentities/NetworkDataSourceWithTotalEntityCountAdapterUnitTest.kt
@@ -5,14 +5,14 @@ import com.xmartlabs.fountain.ListResponse
 import com.xmartlabs.fountain.ListResponseWithEntityCount
 import com.xmartlabs.fountain.Listing
 import com.xmartlabs.fountain.adapter.NetworkDataSourceAdapter
-import com.xmartlabs.fountain.adapter.NetworkDataSourceWithTotalEntityCountAdapter
 import com.xmartlabs.fountain.common.IntMockedListingCreator
-import com.xmartlabs.fountain.common.MockedNetworkDataSourcePageFetcher
-import com.xmartlabs.fountain.common.extensions.generateIntPageResponseList
-import com.xmartlabs.fountain.common.extensions.getPagedList
-import com.xmartlabs.fountain.common.extensions.getPagedListSize
-import com.xmartlabs.fountain.common.extensions.mockLifecycleEvents
-import com.xmartlabs.fountain.common.extensions.sendListResponseWithEntityCountResponse
+import com.xmartlabs.fountain.pagefetcher.NetworkDataSourceWithTotalEntityCountAdapter
+import com.xmartlabs.fountain.testutils.MockedNetworkDataSourcePageFetcher
+import com.xmartlabs.fountain.testutils.extensions.generateIntPageResponseList
+import com.xmartlabs.fountain.testutils.extensions.getPagedList
+import com.xmartlabs.fountain.testutils.extensions.getPagedListSize
+import com.xmartlabs.fountain.testutils.extensions.mockLifecycleEvents
+import com.xmartlabs.fountain.testutils.extensions.sendListResponseWithEntityCountResponse
 import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
