+++ fountain-core/src/test/java/com/xmartlabs/fountain/refresh/RefreshNetworkModeUnitTest.kt
@@ -2,7 +2,7 @@ package com.xmartlabs.fountain.refresh
 
 import com.xmartlabs.fountain.ListResponse
 import com.xmartlabs.fountain.common.IntMockedListingCreator
-import com.xmartlabs.fountain.common.MockedNetworkDataSourceAdapter
+import com.xmartlabs.fountain.testutils.MockedNetworkDataSourceAdapter
 
 class RefreshNetworkModeUnitTest : RefreshUnitTest() {
   override fun createListing(mockedNetworkDataSourceAdapter: MockedNetworkDataSourceAdapter<ListResponse<Int>>) =
