+++ fountain-rx2/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest
-    package="com.xmartlabs.fountain.rx2" />
+<manifest package="com.xmartlabs.fountain.rx2" />
