--- fountain-rx2/src/test/java/com/xmartlabs/fountain/rx2/common/InstantExecutor.kt
@@ -1,8 +1,0 @@
-package com.xmartlabs.fountain.rx2.common
-
-import java.util.concurrent.Executor
-
-
-class InstantExecutor : Executor {
-  override fun execute(runnable: Runnable) = runnable.run()
-}
