--- fountain-rx2/src/test/java/com/xmartlabs/fountain/rx2/common/extensions/IntMockedNetworkDataSourceAdapterExtensions.kt
@@ -1,63 +1,0 @@
-package com.xmartlabs.fountain.rx2.common.extensions
-
-import com.xmartlabs.fountain.ListResponse
-import com.xmartlabs.fountain.ListResponseWithEntityCount
-import com.xmartlabs.fountain.ListResponseWithPageCount
-import com.xmartlabs.fountain.rx2.common.IntMockedListingCreator
-import com.xmartlabs.fountain.rx2.common.MockedNetworkDataSourceAdapter
-import com.xmartlabs.fountain.rx2.common.MockedNetworkDataSourcePageFetcher
-
-fun MockedNetworkDataSourceAdapter<ListResponse<Int>>.sendPageResponse(page: Int = 0) {
-  networkResultListener?.onSuccess(object : ListResponse<Int> {
-    override fun getElements() = generateIntPageResponseList(page)
-  })
-}
-
-fun MockedNetworkDataSourcePageFetcher<ListResponseWithEntityCount<Int>>.sendListResponseWithEntityCountResponse(
-    entityCount: Long,
-    page: Int = 0) {
-  networkResultListener.onSuccess(object : ListResponseWithEntityCount<Int> {
-    override fun getEntityCount() = entityCount
-    override fun getElements() = generateIntPageResponseList(page)
-  })
-}
-
-fun MockedNetworkDataSourcePageFetcher<ListResponseWithEntityCount<Int>>.sendListResponseWithEntityCountResponse(
-    entityCount: Long,
-    elements: List<Int>) {
-  networkResultListener.onSuccess(object : ListResponseWithEntityCount<Int> {
-    override fun getEntityCount() = entityCount
-    override fun getElements() = elements
-  })
-}
-
-fun MockedNetworkDataSourcePageFetcher<ListResponseWithPageCount<Int>>.sendListResponseWithPageCountResponseIfIsRequired(
-    pageCount: Long,
-    page: Int = 0) {
-  networkResultListener.onSuccess(object : ListResponseWithPageCount<Int> {
-    override fun getPageCount() = pageCount
-
-    override fun getElements() = generateIntPageResponseList(page)
-  })
-}
-
-fun MockedNetworkDataSourcePageFetcher<ListResponseWithPageCount<Int>>.sendListResponseWithPageCountResponseIfIsRequired(
-    pageCount: Long,
-    elements: List<Int>) {
-  networkResultListener.onSuccess(object : ListResponseWithPageCount<Int> {
-    override fun getPageCount() = pageCount
-
-    override fun getElements() = elements
-  })
-}
-
-fun generateIntPageResponseList(vararg pages: Int): List<Int> {
-  return pages
-      .distinct()
-      .sorted()
-      .flatMap {
-        val start = it * IntMockedListingCreator.DEFAULT_NETWORK_PAGE_SIZE
-        val end = (it + 1) * IntMockedListingCreator.DEFAULT_NETWORK_PAGE_SIZE - 1
-        (start..end).toList()
-      }
-}
